Update User in schema and database model
Addition of User is_admin and salt attributes to User Model. Updates
made with the intention of adding proper user authentication to access
the pattoo-api. Passwords are now stored as secured hashes along using a
randomly generated salt in which both are stored as attributes in the
pattoo user table.